Added functionality to the Error Modal. Dismissingi it is still a challenge.
Honestly, this one has been a problem, as it is an unfamiliar concept.
I am used only to standard html pop-ups, not those built from scratch.
The challenge here is when I click the dismiss 'Hide' button in the error modal, it dismisss itself.
This it because I am calling event target which is the Button it'seft and not the Modal Component. lol.
To go around this I am adding code to try dismiss the Parrent of the calling component.
Right now I am getting an error here that the component does not have parent property.
It's deeply nested.