Final Code. Fixed some Input value errors by reflecting input to Value property of the input Elements.
Note this I coppied from the Tutor's code. I'd forgotten how to do it. I did change state after submitting the input to the List but it was not reflected back to the Input element.

I have to admit I was already fed up with this project, it was delaying my progress in the module.